move hook in a helper function
