+++ cartridges/custom_cartridge/cartridge/controllers/CheckoutServices.js
@@ -8,6 +8,8 @@ var server = require('server');
 server.extend(module.superModule);
 
 var csrfProtection = require('*/cartridge/scripts/middleware/csrf');
+var HookManager = require('dw/system/HookMgr');
+
 
 /**
  *  Handle Ajax payment (and billing) form submit
@@ -48,7 +50,6 @@ server.replace(
     csrfProtection.validateAjaxRequest,
     function (req, res, next) {
         var PaymentManager = require('dw/order/PaymentMgr');
-        var HookManager = require('dw/system/HookMgr');
         var Resource = require('dw/web/Resource');
         var COHelpers = require('*/cartridge/scripts/checkout/checkoutHelpers');
 
@@ -173,7 +174,6 @@ server.replace(
                 }
             } else {
                 var customerEmail = currentBasket.customerEmail
-                var HookManager = require('dw/system/HookMgr');
                 if (HookManager.hasHook('app.register.requestCustomerToExternalService')) {
                     result = HookManager.callHook(
                         'app.register.requestCustomerToExternalService',
@@ -257,8 +257,9 @@ server.replace(
                 ));
             }
 
-            if (HookMgr.hasHook('app.payment.processor.' + processor.ID.toLowerCase())) {
-                result = HookMgr.callHook('app.payment.processor.' + processor.ID.toLowerCase(),
+            
+            if (HookManager.hasHook('app.payment.processor.' + processor.ID.toLowerCase())) {
+                result = HookManager.callHook('app.payment.processor.' + processor.ID.toLowerCase(),
                     'Handle',
                     currentBasket,
                     billingData.paymentInformation,
@@ -266,7 +267,7 @@ server.replace(
                     req
                 );
             } else {
-                result = HookMgr.callHook('app.payment.processor.default', 'Handle');
+                result = HookManager.callHook('app.payment.processor.default', 'Handle');
             }
 
             // need to invalidate credit card fields
@@ -282,15 +283,15 @@ server.replace(
                 return;
             }
 
-            if (HookMgr.hasHook('app.payment.form.processor.' + processor.ID.toLowerCase())) {
-                HookMgr.callHook('app.payment.form.processor.' + processor.ID.toLowerCase(),
+            if (HookManager.hasHook('app.payment.form.processor.' + processor.ID.toLowerCase())) {
+                HookManager.callHook('app.payment.form.processor.' + processor.ID.toLowerCase(),
                     'savePaymentInformation',
                     req,
                     currentBasket,
                     billingData
                 );
             } else {
-                HookMgr.callHook('app.payment.form.processor.default', 'savePaymentInformation');
+                HookManager.callHook('app.payment.form.processor.default', 'savePaymentInformation');
             }
 
             // Calculate the basket
@@ -348,245 +349,6 @@ server.replace(
         return next();
     }
 );
-
-/**
- * CheckoutServices-PlaceOrder : The CheckoutServices-PlaceOrder endpoint places the order
- * @name Base/CheckoutServices-PlaceOrder
- * @function
- * @memberof CheckoutServices
- * @param {middleware} - server.middleware.https
- * @param {category} - sensitive
- * @param {returns} - json
- * @param {serverfunction} - post
- */
-server.replace('PlaceOrder', server.middleware.https, function (req, res, next) {
-    var BasketMgr = require('dw/order/BasketMgr');
-    var OrderMgr = require('dw/order/OrderMgr');
-    var Resource = require('dw/web/Resource');
-    var Transaction = require('dw/system/Transaction');
-    var URLUtils = require('dw/web/URLUtils');
-    var basketCalculationHelpers = require('*/cartridge/scripts/helpers/basketCalculationHelpers');
-    var hooksHelper = require('*/cartridge/scripts/helpers/hooks');
-    var COHelpers = require('*/cartridge/scripts/checkout/checkoutHelpers');
-    var validationHelpers = require('*/cartridge/scripts/helpers/basketValidationHelpers');
-    var addressHelpers = require('*/cartridge/scripts/helpers/addressHelpers');
-
-    var currentBasket = BasketMgr.getCurrentBasket();
-
-    if (!currentBasket) {
-        res.json({
-            error: true,
-            cartError: true,
-            fieldErrors: [],
-            serverErrors: [],
-            redirectUrl: URLUtils.url('Cart-Show').toString()
-        });
-        return next();
-    }
-
-    var validatedProducts = validationHelpers.validateProducts(currentBasket);
-    if (validatedProducts.error) {
-        res.json({
-            error: true,
-            cartError: true,
-            fieldErrors: [],
-            serverErrors: [],
-            redirectUrl: URLUtils.url('Cart-Show').toString()
-        });
-        return next();
-    }
-
-    if (req.session.privacyCache.get('fraudDetectionStatus')) {
-        res.json({
-            error: true,
-            cartError: true,
-            redirectUrl: URLUtils.url('Error-ErrorCode', 'err', '01').toString(),
-            errorMessage: Resource.msg('error.technical', 'checkout', null)
-        });
-
-        return next();
-    }
-
-    var validationOrderStatus = hooksHelper('app.validate.order', 'validateOrder', currentBasket, require('*/cartridge/scripts/hooks/validateOrder').validateOrder);
-    if (validationOrderStatus.error) {
-        res.json({
-            error: true,
-            errorMessage: validationOrderStatus.message
-        });
-        return next();
-    }
-
-    // Check to make sure there is a shipping address
-    if (currentBasket.defaultShipment.shippingAddress === null) {
-        res.json({
-            error: true,
-            errorStage: {
-                stage: 'shipping',
-                step: 'address'
-            },
-            errorMessage: Resource.msg('error.no.shipping.address', 'checkout', null)
-        });
-        return next();
-    }
-
-    // Check to make sure billing address exists
-    if (!currentBasket.billingAddress) {
-        res.json({
-            error: true,
-            errorStage: {
-                stage: 'payment',
-                step: 'billingAddress'
-            },
-            errorMessage: Resource.msg('error.no.billing.address', 'checkout', null)
-        });
-        return next();
-    }
-
-    // Calculate the basket
-    Transaction.wrap(function () {
-        basketCalculationHelpers.calculateTotals(currentBasket);
-    });
-
-    // Re-validates existing payment instruments
-    var validPayment = COHelpers.validatePayment(req, currentBasket);
-    if (validPayment.error) {
-        res.json({
-            error: true,
-            errorStage: {
-                stage: 'payment',
-                step: 'paymentInstrument'
-            },
-            errorMessage: Resource.msg('error.payment.not.valid', 'checkout', null)
-        });
-        return next();
-    }
-
-    // Re-calculate the payments.
-    var calculatedPaymentTransactionTotal = COHelpers.calculatePaymentTransaction(currentBasket);
-    if (calculatedPaymentTransactionTotal.error) {
-        res.json({
-            error: true,
-            errorMessage: Resource.msg('error.technical', 'checkout', null)
-        });
-        return next();
-    }
-
-    // Creates a new order.
-    var order = COHelpers.createOrder(currentBasket);
-    if (!order) {
-        res.json({
-            error: true,
-            errorMessage: Resource.msg('error.technical', 'checkout', null)
-        });
-        return next();
-    }
-
-    // Handles payment authorization
-    var handlePaymentResult = COHelpers.handlePayments(order, order.orderNo);
-
-    // Handle custom processing post authorization
-    var options = {
-        req: req,
-        res: res
-    };
-    var postAuthCustomizations = hooksHelper('app.post.auth', 'postAuthorization', handlePaymentResult, order, options, require('*/cartridge/scripts/hooks/postAuthorizationHandling').postAuthorization);
-    if (postAuthCustomizations && Object.prototype.hasOwnProperty.call(postAuthCustomizations, 'error')) {
-        res.json(postAuthCustomizations);
-        return next();
-    }
-
-    if (handlePaymentResult.error) {
-        res.json({
-            error: true,
-            errorMessage: Resource.msg('error.technical', 'checkout', null)
-        });
-        return next();
-    }
-
-    var fraudDetectionStatus = hooksHelper('app.fraud.detection', 'fraudDetection', currentBasket, require('*/cartridge/scripts/hooks/fraudDetection').fraudDetection);
-    if (fraudDetectionStatus.status === 'fail') {
-        Transaction.wrap(function () { OrderMgr.failOrder(order, true); });
-
-        // fraud detection failed
-        req.session.privacyCache.set('fraudDetectionStatus', true);
-
-        res.json({
-            error: true,
-            cartError: true,
-            redirectUrl: URLUtils.url('Error-ErrorCode', 'err', fraudDetectionStatus.errorCode).toString(),
-            errorMessage: Resource.msg('error.technical', 'checkout', null)
-        });
-
-        return next();
-    }
-
-    // Places the order
-    var placeOrderResult = COHelpers.placeOrder(order, fraudDetectionStatus);
-    if (placeOrderResult.error) {
-        res.json({
-            error: true,
-            errorMessage: Resource.msg('error.technical', 'checkout', null)
-        });
-        return next();
-    }
-
-    if (req.currentCustomer.addressBook) {
-
-        // save all used shipping addresses to address book of the logged in customer
-        var allAddresses = addressHelpers.gatherShippingAddresses(order);
-        allAddresses.forEach(function (address) {
-            if (!addressHelpers.checkIfAddressStored(address, req.currentCustomer.addressBook.addresses)) {
-
-                // save all used shipping addresses to address book of external cervice
-                var UUIDUtils = require('dw/util/UUIDUtils');
-                var integrationAddressId = UUIDUtils.createUUID();
-                var CustomerMgr = require('dw/customer/CustomerMgr');
-                var customer = CustomerMgr.getCustomerByCustomerNumber(
-                    req.currentCustomer.profile.customerNo
-                );
-                var profile = customer.getProfile();
-                var result;
-
-                var HookManager = require('dw/system/HookMgr');
-                if (HookManager.hasHook('app.register.requestCustomerToExternalService')) {
-                    result = HookManager.callHook(
-                        'app.register.requestCustomerToExternalService',
-                        'addCustomersAddressToExternalService',
-                        address,
-                        integrationAddressId,
-                        profile.custom.v_integrateId
-                    );
-                }
-                if (!result.ok) {
-                    return;
-                }
-                //add integration id to the new created address
-                address.integrateAddressId = integrationAddressId;
-
-                addressHelpers.saveAddress(address, req.currentCustomer, addressHelpers.generateAddressName(address));
-            }
-        });
-    }
-
-    if (order.getCustomerEmail()) {
-        COHelpers.sendConfirmationEmail(order, req.locale.id);
-    }
-
-    // Reset usingMultiShip after successful Order placement
-    req.session.privacyCache.set('usingMultiShipping', false);
-
-    // TODO: Exposing a direct route to an Order, without at least encoding the orderID
-    //  is a serious PII violation.  It enables looking up every customers orders, one at a
-    //  time.
-    res.json({
-        error: false,
-        orderID: order.orderNo,
-        orderToken: order.orderToken,
-        continueUrl: URLUtils.url('Order-Confirm').toString()
-    });
-
-    return next();
-});
 
 
 module.exports = server.exports();
